Fix various warning around conversion of types

+++ tutorial/detection/tag/tutorial-apriltag-detector-live.cpp
@@ -215,7 +215,7 @@ int main(int argc, const char **argv)
   double opt_tag_size = 0.053;
   float opt_tag_quad_decimate = 1.0;
   float opt_tag_decision_margin_threshold = 50;
-  float opt_tag_hamming_distance_threshold = 2;
+  int opt_tag_hamming_distance_threshold = 2;
   int opt_tag_nThreads = 1;
   std::string intrinsic_file = "";
   std::string camera_name = "";
